purple-guardian-playwright/requirements.txt
playwright>=1.40.0
typing-extensions>=4.0.0
pydantic>=2.0.0
colorama>=0.4.6
rich>=13.0.0

# Optional dependencies for full functionality
PyYAML>=6.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0